requirements.txt
open-interpreter
fastapi
uvicorn
python-dotenv
